Remove unnecessary variable assignments, move errorHelper to src/constants
