+++ src/helpers/authMiddleware.js
@@ -1,14 +1,14 @@
-const errorHelper = require('../helpers/errorHelper.js');
+const errorMessages = require('../constants/errorMessages.js');
 const yotiHelper = require('../helpers/yotiHelper.js');
 
 module.exports = function (req, reply) {
   let token = req.auth.credentials.auth;
   if(!token) {
     return reply.view('error', {
-      error : errorHelper.loginError
+      error : errorMessages.loginError
     });
   }
-  let user = yotiHelper(token, function(err, user) {
+  yotiHelper(token, function(err, user) {
     if (err) {
       return reply.view('error', {
         error : err
