+++ src/routes/assign.js
@@ -1,6 +1,6 @@
 const authMiddleware = require('../helpers/authMiddleware.js')
 const requestTable = require('../database/tables/requests.js')
-const errorHelper = require('../helpers/errorHelper.js');
+const errorMessages = require('../constants/errorMessages.js');
 
 module.exports = {
   method: 'GET',
@@ -20,7 +20,7 @@ module.exports = {
       requestTable.updateAssigned(request, function (err, data) {
         if (err) {
           return reply.view('error', {
-            error: errorHelper.databaseError
+            error: errorMessages.databaseError
           });
         }
         reply.redirect('/dashboard');
