+++ src/routes/submit.js
@@ -1,6 +1,6 @@
 const authMiddleware = require('../helpers/authMiddleware.js');
 const requestTable = require('../database/tables/requests');
-const errorHelper = require('../helpers/errorHelper.js');
+const errorMessages = require('../constants/errorMessages.js');
 
 module.exports = {
   method: 'POST',
@@ -16,7 +16,7 @@ module.exports = {
       requestTable.insert(req.payload, req.pre.user, (err) => {
         if(err) {
           return reply.view('error', {
-            error : errorHelper.databaseError
+            error : errorMessages.databaseError
           });
         }
         reply.view('thankyou');
